backend/requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
psycopg2-binary>=2.9.5,<2.10.0
python-dotenv>=1.0.0,<1.1.0
gunicorn>=20.1.0,<20.2.0
drf-yasg>=1.21.7,<1.22.0
django-filter>=23.3,<23.4